15652 kube_cron_job_status: introduction of execution duration metric
Prior to this werk, the check only displayed the total duration of the job
which included the pending time duration alongside the actual running
duration. This werk introduces the 'Execution duration' metric which is solely
based upon the actual running duration. To avoid confusion the previous graph
'Duration' is renamed to 'Total Duration'.

CMK-12803

Change-Id: Ibba18b1c8ea8d1bcfba902e11f81612deb900b42