15264 FIX Check plugins logwatch, logwatch_ec: Notify about unreadable log file
In the Logwatch and Logwatch Event Console Forwarding Check unreadable log files (e.g. due to insufficient permssions) were previously treated the same as missing log files.
With this werk, the checks will check specifically for log files that could not be read and will become {CRIT} if any are present.
To configure the state for the Event Console Forwarding service when unreadable files are detected, use the option "State if a logfile cannot be read" in the rule <i>Logwatch Event Console Forwarding</i>.

SUP-12770

Change-Id: I4d0186b76235396b864ceadcdecf032e45fd7e3d