Avoid paths as default params
Default arguments are bound on import and so are class variables.  Local
variables are bound on use.  Now, our paths are patched in the tests and
the tests do not control the imports.  Thus, there is a non-zero chance
that patching `omd_root`, `$OMD_ROOT`, and friends doesn't work as
expected in the tests.

This caused over 500 failing unit tests in an unrelated change.

In short:  if you mutate a constant (here, by changing the root of the
paths) then it's not a constant.  Having mutable default arguments in
Python is asking for trouble.

Change-Id: I6094eb08e1cb17cc24f28aa2f851d8384b4cd027