Build Setup search index at request start (instead of end)
This has the advantage that we can be certain that everything is
correctly loaded. For example, enabling an MKP might add new rulespecs.
If we built the search index at the end of the same request in which the
MKP was enabled, these rulespecs would not be taken into account (since
the Apache reload which happens after enabling the MKP only affects
future requests). Hence, we instead trigger the build at the beginning
of the next request.

CMK-12830

Change-Id: Ia02fc5459b263296ee9f78934c436bfda5b3745d