FIX 15463 mk_filestats: Make combination of file grouping and single file aggregation more usable
This werk is incompatible for users that use file grouping and single file aggregation
and have "%s" in the section name.
It's now possible to define a name containing "%s" in both section and group names when
using single file aggregation.
In that case, the services of files that don't belong to a group will be named
"<SECTION NAME> <FILE NAME>" and the services of grouped files
"<SECTION NAME> <GROUP NAME> <FILE NAME>".

CMK-12778

Change-Id: I7f523851cd5e3214ee087bb2a34962789c5a97f7