FIX 15467 sap_hana: Enable clustering of active and passive nodes
Previously, we collected data only for databases in primary and
logreplay_readaccess replication modes. For other replication modes
only SAP HANA Status service was shown.

In order to enable clustering of active and passive nodes,
we now try to collect data for all replication modes and if the database
cannot be accessed, the services of that host will go to STALE.

SUP-4970

Change-Id: Id3e26260a3707bf5997302ef54a2c72806c7d2d9