15678 FIX schemas: regex pattern updates
This werk introduces a fix to the pattern attribute in various
rest api schemas.

To avoid partial matches, regex patterns should start with
the caret and end with the dollar symbol.

Also, the + symbol should be preferred over the * symbol as
the * symbol allowed for zero or more matches instead of
1 or more which is what the + symbol gives us.

For patterns that list possible options, e.g.
  monday|tuesday|wedneday... etc
they have instead been removed and replaced with enums.

CMK-12220

Change-Id: I89dab0652c15cb9e7381e1f6d5d974afaf0bfd40