Units: Introduce optional field for conversion
Since converting when doing the actual rendering is too late, we instead
need a function to convert the actual data points.

Also remove the current conversion mechanism for the temperature unit
and add the new field instead. As of this commit, the conversion of the
temperature unit is temporarily broken, until we apply the conversion to
the various places where data is retrieved from livestatus.

CMK-12958

Change-Id: Ie1c67d8c9064e47754ec739ba46ebd5c7aafb9f9