13982 SEC Reading host_config's will now honour contact groups
Prior to this Werk it was possible for a user to read a hosts configuration
(using GET on '/objects/host_config/<host_name>') even if that user was not
in the contact group of that host.

The REST-API will correctly check a users permissions before serving a response
in that case and report a 403 error if the user cannot access the host's config.

<b>Affected Versions</b>:
LI: 2.2.0 (beta)
LI: 2.1.0

<b>Vulnerability Management</b>:
We calculated a CVSS 3.1 score of 4.3 (Medium) with the following vector: CVSS:3.1/AV:N/AC:L/PR:L/UI:N/S:U/C:L/I:N/A:N
We assigned CVE-2023-22348 to this vulnerability.

We found this vulnerability internally and have no indication of any exploitation.

Change-Id: Id04281db696467ae11ee1d5ce3d172c1bed71a93