15559 FIX mk_oracle: Always use TNS_ALIAS when present
This werk may be relevant for you, in case:

<ul>
<li>you want to monitor remote instances</li>
<li>you are using a TNS_ALIAS to reach the remote instance</li>
</ul>

Previously, a <tt>TNS_ALIAS</tt> from mk_oracle.cfg was pinged by the plugin with <tt>tnsping</tt> in order to verify that this alias can be reached.
However, this needs the tool <tt>tnsping</tt> installed with the oracle client, which is not shipped per default.

A failure led to silently ignoring the <tt>TNS_ALIAS</tt> and falling back to a potentially invalid connection string which is passed to <tt>sqlplus</tt>.
The plugin now assumes that the given <tt>TNS_ALIAS</tt> can be reached and uses it.

Change-Id: I3d29b86d1071159d5371aff191f01704e9f54226
JIRA-Ref: SUP-12962