Historical & forecast metric painters: Simplify rendering
The rendering was unnecessarily complicated: After computing the final
value to be rendered, including metric translations and unit
conversions, this value was transformed into a perfomance data string.
This string was subsequently parsed to extract the numerical value. Not
only is this unnecessary, it would also soon lead to applying unit
conversions twice: Once after fetching the historical data and once
after parsing the artificially created performance data string.

CMK-12958

Change-Id: I578a5427924dc71228795c4410b007e3c6df5d51