15679 FIX host_config: moving a host to a different folder only permitted if user has permission
This werk addresses an issue that allowed an api user to move a host from one
folder to another when the user didn't belong to the same contact group. The api
user now has to be part of the same contact group as the source folder, target
folder and the host in order to be able to move the host. Please update any client
scripts accordingly.

CMK-10569

Change-Id: Ib50b1f4c308d887ca605aedde65e91c89c9ab25b