15266 FIX Check plugin logwatch_ec: Improve detection of missing log files
Previously the Logwatch Event Console Forwarding service would not report log files as missing that are marked as such in the agent output:
F+:
<<<logwatch>>>
[[[logfile:missing]]]
F-:
This affected logfiles that were explicitly configured to be monitored, as opposed to configurations that capture multiple logfiles at once.
With this werk this has been fixed.

SUP-12770

Change-Id: Idfc972fce51c0bdeaaf7f73d56d38d15e59cdb34