14308 FIX Fix "Invalid job status None found" message for background jobs
In case a job has just started and did not report any status update or was
interrupted during startup the `jobstatus.mk` state file could be empty,
resulting in the message Invalid job status None found in [PATH]".

The handling of this situation has been improved by showing the initialized
state in this situation.

Change-Id: Ic82b06661e12eab42b6a1d6bf4498652b557d56a