15680 FIX downtimes: duration field changed from seconds to minutes
This werk addresses an issue with the downtimes endpoint.  When
creating a downtime, the duration field was accepting an integer
that represented the downtime in seconds. The problem was that
the gui shows this in minutes. The fix changes the duration
field to accpet integers that represent minutes instead.

CMK-7183

Change-Id: Ie8a7e30a8fb950b343e7ce77fb7f7909650f871d