kube: introduce requests session as api_client I
This commit prepares the making the calls to the Kubernetes API server
asynchronous.  The Kubernetes Python client does offer an async option
itself. With this option, the HTTP requests not be executed in the main
thread. However, they will still be blocking each other.

Moreover, the client cannot be pickled, thus multiprocessing is also not
possible.

Therefore, we now move to a different option. This change modifies the
smallest of the API classes. More updates will follow.

CMK-12359

Change-Id: Ia111cb8d320257162823aed3604657d19b6da9c2