kube: introduce requests session as api_client II
This commit prepares the making the calls to the Kubernetes API server
asynchronous.  The Kubernetes Python client does offer an async option
itself. With this option, the HTTP requests not be executed in the main
thread. However, they will still be blocking each other.

Moreover, the client cannot be pickled, thus multiprocessing is also not
possible.

Therefore, we now move to a different option. This change modifies the
CoreAPI calls, which did not use the client for deserialization.

CMK-12359

Change-Id: Ieac2633aa7db94e53ab89d3cd16369409a732974