kube: introduce requests session as api_client III
This commit prepares the making the calls to the Kubernetes API server
asynchronous.  The Kubernetes Python client does offer an async option
itself. With this option, the HTTP requests not be executed in the main
thread. However, they will still be blocking each other.

Moreover, the client cannot be pickled, thus multiprocessing is also not
possible.

Therefore, we now move to a different option. This change completes the
transition. It also introduces a new way of calling the Kubernetes
client to still perform the deserialization.

CMK-12359

Change-Id: I0cd8f14f4d6c79b6abb602533d57fb0ccd5445f7