FIX 15466 autodiscovery: Show site changes made by Periodic service discovery
The werk is incompatible because it removes the 'cmk --discover-marked-hosts' command.

Previously, if 'Periodic service discovery' was used with 'Automatically update service configuration'
enabled and 'Do not activate changes' activation option, the services would be discovered but
the change wasn't visible in 'Activate pending changes' or in the audit log.

This made it impossible to know whether there are any discovered changes that need to be activated
and what these changes are.

Now, when changes get automatically discovered but not activated, they appear in
'Activate pending changes' and the action is logged to the audit log.
In case of automatic discovery with activation, the action will be logged to the audit log.

SUP-9139

Change-Id: I5c20708b22665a68b0f8e063912772c24845690b