do not read cluster host labels file
We will remove the clusters host label files in 2.3,
and always compute the labels on demand.
This is consistent with the service discovery.

Since the files are synced to the central site,
(which might still need them) we still *write* the files.

Note that the adaptions in the tests are either fixes
(we do not expect "vanished" labels if we did not read
the existing ones) or results of an inconsistent test
scenario where clusters had labels that did not arise
from one of the nodes.

Change-Id: If43e8939b711db1f7308f266a6b56f1c32635be1