do not read cluster host labels file
We always compute the labels on demand.
This is consistent with the service discovery.

Note that the adaptions in the tests are either fixes
(we do not expect "vanished" labels if we did not read
the existing ones) or results of an inconsistent test
scenario where clusters had labels that did not arise
from one of the nodes.

Change-Id: If43e8939b711db1f7308f266a6b56f1c32635be1