Fix backup_lock_dir_fixture
The sudo calls the fixture tried to make failed. This made both cases
of the fixture work with a pre-existing `/var/run/mkbackup` directory.

The issue was undetected because we did not check the exit code of the
command. The output of the failure was only visible with `pytest -s`.

Change-Id: I7697469c5c2441a911bd5609ffce5dc63ee69dec