Stop using allocated_ports file
This file is unnecessary and it was buggy. For example, calling `omd
status` reset the file contents to default values (due to the config
loading hooks such as APACHE_TCP_PORT which are executed during omd
commands). Afterwards, the file was out of sync with site.conf. Creating
a new site could then lead to ports being allocated twice.

FEED-7861

Change-Id: I2e8b27391116247fbc52f35edfa3a7adf07156cf