15728 FIX mk_oracle: revert werk 15210
The introduction of werk 15210 brought convenience for some, while
it unintentionally broke the rules for others. This new werk reverts
the rule set back to the previous behaviour. Users should reverify
the auth values for host and port in their existing rules and
disable those options if this was previously the case.

Change-Id: I81b17efb1f9fb79948f02cef7f01a2753bfc1ed5