Move announce texts from bauwelt to checkmk
Before bauwelt (in bw-release and bw-get-changes) directly called
functions defined in cmk.utils.werks to build the announcements.

Now the creation of announcement is moved into the checkmk repository.

This way the interface between bw-release is no longer the python code.
bw-release will be responsible to download announce*.tar.gz create with

  make `pwd`/announce-2.3.0b1.tar.gz

in the build process and load the announcement text from there.

The text is generated twice: One for the announcement for the forum in
markdown, and one in plain text for the mailing list.

Change-Id: I5c272f823cf5169bb12cc8fbc0d75e3c20339d60