Cleanup Folder.save
The save method now does not need to call load_instance anymore.
It would also be nice to remove the invalidate_caches in this method,
since it is kind of suprising that a save methods triggers it.
However, removing it makes some test fail which needs to be investigated
later.

Change-Id: If9b623e100e77d8a8cebec6dafec9279b1e0c359