15847 FIX <tt>cisco_qos</tt>: Handle metric units consistently & drop averaged metrics
cisco_qos now uses bits internally, which is easier, since we want to
store our metrics in this unit.

Explanation regarding the CRE issue mentioned in the werk:
Livestatus in combination with pnp4nagios only reports metrics which are
currently still updated. Metrics which were once produced but are not
updated anylonger are currently not reported in the livestatus metrics
column. Hence, renaming metrics currently leads to a loss of historic
data in the CRE, even if there is a corresponding translation.

Change-Id: I605587c4b72651a96bd18eba07cd4191797056eb