remove optional import of graphs.ts in mobile.ts
This file used to be a part of CEE before and thus were optionally
imported in case it's not available in raw version. However, now it's
a part of the raw version so we don't need to optionally import it,
since it's alaways available. I also changed the license of graphs.ts
and _graphs.scss to match the new situation.

Change-Id: Iea99b181685e2e15f08ed20ba72ee353ad9147b6