Extract collection related staticmethod from Folder
The folders, used to organize hosts, are represented in a tree. So far
the logic for interfacing with the tree was implemented as staticmethods
on the Folder class. Together with all the other code in the Folder
class this created a hard to understand mess.

This change now extracts these functions in order to separate the code
for managing a tree of folders.

It's still far from ideal. The state was previously held in a request
global which we also need to do now to not end up with a huge change. In
subsequent steps it might make sense to propagate the FolderTree object
through the application. However, the goal for now is to simplify the
Folder class by extracting functionality from it.

Change-Id: Ie6651aaede27f02e87532a41680320c1ec5f4ac0