15654 windows_tasks: allow discovery to decide if disabled tasks are included or not
The Windows Tasks rule allowed the user to map not 'enabled' tasks to a specific state.
However, tasks which were 'disabled' during discovery were completely ignored and
could therefore not benefit from this rule. This led to confusion as to what group was
targeted by the specification 'not enabled'. This werk, therefore, introduces an appropriate
discovery rule which allows the user to control if initially 'disabled' tasks are supposed
to be monitored or not.

SUP-14018

Change-Id: I9bab674a57b2825a4195f678aa0281c565c801b9