15268 FIX Incorrect edition provided by livestatus after update using Nagios core
This werk is relevant to users operating Checkmk with Nagios core that are using the new "edition" livestatus column or the Checkmk Cloud Edition with Distributed Monitoring.
If such a system is updated to another edition, the livestatus information is not updated with it.
Since this information is used to determine compatibility between central and remote sites in Distributed Monitoring, sites can incorrectly be detected as incompatible.

With this werk, an update will update the livestatus information as well.

CMK-13413

Change-Id: I7c222e4ff3327cddcad1578ae0df8b80298b8f9b