15753 FIX agent_azure: Make resource group filter case-insensitive
Due to the inconsistant behaviour of the Azure API, the filtering of
resources by resource group wasn't working as expected.
The resource group name returned by the API was sometimes of a different
case then the one visible in the Azure portal, which led to resources not
being discovered by Checkmk.

Now, the resource group filtering is case-insensitive and the problem
no longer occurrs.

CMK-13026

Change-Id: If6d0c596a0fc7271cc157d574982d5781cfa8f31