15267 FIX License verification: Incorrect error message when saving licensing credentials
This werk is relevant for users that use the Checkmk Cloud Edition and use the Online Verification modus to verify their license.
If the licensing credentials are configured under <tt>Setup > Maintenance > Licensing > Edit settings</tt> and the settings are saved, an error message will be shown.
The message claims that the last verification failed with the specific error "Verification failed (The Licensing ID or password you entered is incorrect.)". In this instance, the message is incorrect and the verification can be attempted regardless.

With this werk this has been fixed.

Change-Id: I1124f60a753728c1b41c3d31f1fb14eccbf53b18