Re-work functionality for handling rrd data relevant for a single metric
The current implementation is conceptually wrong: Currently, we scale
the rrd data right when fetching it via livestatus. This is wrong
because we can only know if we should apply any scalings in the context
of a service (since our metric translations are check-plugin specific).

Example: Let's say we want historic values for the metric mem_total.
Outside the service context, we cannot know whether or not we should
scale this metric by 1024**2 (Windows: yes, Linux: no).

This issue affects certain painters and the single-metric dashlets.
(though most of the latter, namely the ones where the host and service
are fixed, could be re-implemented to completely avoid this issue, which
is however outside the scope of this change). Note that eg. service
graphs are not affected, because we are in the context of a service when
rendering them.

This change leads to some duplicated code, however, this simplifies the
review. In the following commit, we will switch from the old to the new
functionality and then remove the old code.

CMK-13561

Change-Id: I0e177242d58c4d39e268db37c4c42df37bfec9a9