15755 ps: Make showing usernames in service details configurable
An additional option "Include usernames in service details" is added to
the "State and count of processes" service monitoring rule.
It enables the user to configure whether the usernames are shown in
the service details of the "Process" service.

Usernames will be shown in the service details by default.

CMK-13118

Change-Id: I237ada9e47dde9a70949d8120fa26b1587b4a9ce