15879 FIX host_config: introduction of background job for host_rename endpoint
Prior to this werk, the host_rename endpoint occasionally ran into a timeout if
the renaming process took too long. This werk introduces a background job
to resolve this issue. The endpoint now returns a redirect response instead of
the host details. An associating wait-for-completion endpoint is also introduced
to accompany the rename endpoint. Scripts should be adjusted accordingly.

Change-Id: I2febe0fbee1e806a476a6a9fbf76e8eff0b53a97