Add factories to load and create folders
The goal is to extract the alternative construction of new folders and
loading of existing folders from the constructor which will then help to
reduce the arguments of both variants. This is a first step which keeps
the arguments of the constructor. Next step will be to move over the
code to the specific places and reduce the arguments to the required
minimum.

Change-Id: I02aab2b63ba19c965b427583a78580e8c2f7ad04