FIX: State and Count of Processes: Calculate process creation time consistently
The creation time of a process, as displayed by the "State and Count of Processes" service is calculated by taking the difference between the current time and the time elapsed since the process was started.
Previously, this used the time the check was performed at as the current time, leading to different creation times if different services were created for the same process.
Now the time the agent section was parsed is used instead.

Note that this will still lead to incorrect creation times if it takes too long to fetch the agent output.

SUP-14394

Change-Id: If282dc8258395a69a67344b448fdfec86d0c2f6f