15896 FIX State and Count of Processes: Use remote server time for process creation time when using Solaris or AIX agent
As mentioned in werk 15894, the reference time used to calculate the creation time of a process as displayed by the "State and Count of Processes" service is the time the agent section was parsed.
The <tt>check_mk_agent.solaris</tt> and <tt>check_mk_agent.aix</tt> agents will now also output the time at which the ps command was executed to be used as reference instead.
The agent must be redeployed to take advantage of this change.

SUP-14394

Change-Id: I734ef73b6707310543b7e8bc65160a095742d867