15905 FIX check_disk_smb: Support SMB share hosts in networks other than the Checkmk server network
With this werk check_disk_smb now supports share hosts which are in a different network than the Checkmk server. To enable this, the corresponding ruleset now offers an option for configuring the IP address.
Note that this option is only needed if the HOST is not in the same network as the Checkmk server.
Before this werk, if the SMB share host was in a different network, the plugin crashed.

Change-Id: Ie27af6732726a797fb1b1987abb41376adc2951e