15759 FIX mk_oracle: Fix error in the mk_oracle plugin
Werk 15848 (version 2.1.0p29) introduced a bug which caused the mk_oracle
plugin to execute incorrectly and stop sending data to Checkmk.

If called from the command line from the plugin's parent directory, the plugin
incorrectly reported `Unhandled location` and terminated, even if the location
of the plugin was correct.

SUP-14507

Change-Id: I2d3b4b4e8eef3bd7ca3709efd995ef6b245f5128