Move `HTTPMethod` to gui
Actually, the type is only used in the openapi but moving it
there either creates dependency cycles or breaks 4000+ tests.

Anyway, `cmk.gui.http` is still an improvement from `cmk.utils`
as there is, to my knowledge, no HTTP-related code in utils or
base so this change still improves the cohesion.

Change-Id: I42ebe89c4b7b2c0a8891d3020475735155575fe0