Move root_dir to FolderTree
This feature fits better with the FolderTree class instead of the
CREFolder. The CREFolder only needs to be aware of it's path relative
to the root provided by the FolderTree.

To enable this change the tree needs to be made available to various
helpers, like _RedisHelper and FolderMetadata now. This also eliminates
some implicit dependencies on global states.

Change-Id: I316076e2fdf2a9309f47b858229e0fc6ad88aeef