15760 FIX mk_oracle: '--no-spool' command-line option added
Previously, mk_oracle behaved differently depending on if it was
run from CLI or the Checkmk agent. The approach wasn't consistent
and led to errors if the script was called from CLI without arguments.

Now, mk_oracle will always behave as if run from the agent unless
additional arguments are provided.

Additionally, new command-line argument was introduced, '--no-spool'.
If used on host with systemd, the output of mk_oracle will be written
to stdout instead of the spool file. The option doesn't influence the behavior
of the script if xinetd is used.

SUP-14521

Change-Id: Ibfd1f1b4d874ebbe27c40869b808394be699f5b9