Increase ulimit to avoid bazel crash
When building Python utilizing the Bazel cache we got the following
error message:
`Server terminated abruptly (error code: 14, error message: 'Connection
reset by peer'...`
We can reproduce this error more reliably by decreasing the ulimit,
while we are not seing this error with increased ulimit.
The ulimit was set to 1024 because of Python2: KNW-141
Since we are using Python3 to it seems safe to increase the ulimit.
As a positive side effect we see an increase in build speed.

Change-Id: I3c878f12e68a706390bf8cdef6ab3ec8fa3f0949