15767 FIX check_mk_agent: Fix crash in AIX and Solaris agents if agent profiling is on
AIX and Solaris agents would previously crash if agent profiling was activated
with LOG_SECTION_TIME=true.
The error in the agent output was:
C+:
./mk_filestats.py.log: cannot create [No such file or directory]
C-:

tr command was replaced with sed because it didn't work properly on some Solaris
and AIX environments. It left spaces in the name of the log file.

SUP-14208

Change-Id: I9310c30d620a96dd9a240c407d919e52e7ae5f56