15881 saml2: addition of option to map attribute value to specific contact groups
Prior to this werk, the contactgroup attribute value had to match an existing
Checkmk contact group name in order for the user to be assigned to the respective
group. This werk introduces the option for the attribute value to be mapped to
multiple contact groups in case of a name match.

Change-Id: I663f2cb127e10d50aa36a33a16597776faf7785e