FIX 15765 check_temperature: Fix misleading service details in temperature checks
In case the state from the device wasn't OK, thresholds from the device were
appended to the service details, e.g. '46 °C, warning (device warn/crit at 56/56 °C)'.
This suggests that the temperature value exceeds the threshold, which isn't the case.

To make it clearer, the service details are changed to: '46 °C, State on device: warning'
in this case.

SUP-14504

Change-Id: I47011819bc78c58860fa3b81a9735bb317b1c8a7