Fixes:  wrong order in the output of the metrics in the interface bandwith gaph
The interface bandwidth graph displays the input/output bandwidth metrics in a non-uniform order. In the diagram, the input bandwidth is in the upper half and the output bandwidth is in the lower half of the diagram. In the legend and tooltip, it is the other way around.

This PR will change the order of the Input/Output bandwith in the legend and the tool tip.

For more (graphical) detail have a look at 

https://forum.checkmk.com/t/servicegraphen-kosmetische-anfrage/29906/8