15769 FIX oracle_performance_memory: Check all memory metrics
Previously, it was possible to configure thresholds for all memory
metrics, but only the "Maximum SGA Size" metric was checked. This resulted
in an incorrect "ORA Performance Memory" service state.

This is already fixed in newer branches with a migration of
oracle_performance checks to the new Check API.

SUP-14560

Change-Id: I7bf4772e3512410b1e5375ea2d7e0cb3a438fabc