Prevent previous version .pkl files from breaking cmk-update-config
The hosts.pkl files exist to improve read performance on the hosts.mk
files. Similar to what .pyc does for .py files.

There may be references to non native types in these files. For example,
in 2.2 we store a reference to the UserId with it's absolute module path.
In the moment we move or rename this type, the .pkl files can not be
loaded anymore.

This means we have no guarantee that the files can be read by the
cmk-update-config code of the version a user just updated to.

The current concept of the .pkl files allows removal of the files, since
they can be regenerated from the .mk files.

So we remove the .pkl files in an early stage of the update procedure
from now on.

CMK-13841

Change-Id: I91cff3d8d1ff2c5243bdce92a08c83e1efcadda7