Fixes: is not a valid PhysicalClasses
In PhysicalClasses are only the values from 1 to 12 defined (this acording to the MIB description).
For all other values is an explicit errorhandling implemented (currently only for "0","" and "14").
This fix will change this, so that for every value other than 1-12 the physical classe "unknown" is returned.

Werk 15633 was only fixing ths error for the value 14.

One Example from a Cisco Stack:

        ['1001', '13', 'StackPort1'],
        ['1002', '14', 'StackAdapter1'],

Here the value 13 results in a crash of the inventory plugins.

Change-Id: I4f59caf1f5899eb726dc8e86d1435df00610b68a