15550 FIX Grouped services crash and incorrect visualization
Using grouped services containing unescaped regex chars, like <b>\P</b>,
in their title would cause views displaying these services in a group to crash.

Now the views are displayed as expected.

Also the visualization of group title is now fixed. The content of a match group is only shown in the title of the
grouped services if a match group is actually being used in grouping.

SUP-14585

Change-Id: I9bb7d18b746f198da3e15aef5e603a7eaae2b196