Purge WithAttributes mixin
The accessors prevented easy typing and did not add enough value.
So we hold the attributes now as a member of CREFolder, CREHost and
SearchFolder for the sake of simplicity.

Had to weaken the attributes typing at some places to be able to
merge it. It's temporarily acceptable since the follow up commit
will change the type to a TypedDict, which will make it work again
in a better way.

Change-Id: I23d92a3d43b761a21939afe7c47ec6633c10040a