15969 FIX cisco_temperature: Correct warning threshold for environment temperature
Currently only the threshold specified by OID 1.3.6.1.4.1.9.9.13.1.3.1.4 (ciscoEnvMonTemperatureThreshold) is considered when checking the temperature of a Cisco switch. As only one threshold is given, it is used as both WARN and CRIT threshold, causing confusion when the device itself reports a warning state based on a previously unknown WARN threshold.
Now the upper thresholds specified in OID 1.3.6.1.4.1.9.9.91.1.2.1.1 (entSensorThresholdTable) will be used as WARN and CRIT thresholds if an entry for the relevant switch can be found.
For this, the threshold relation and severity will be used to determine if the threshold can be applied in Checkmk, otherwise it will be ignored.

SUP-14653

Change-Id: If8bc63d378d3a77fbf94b5ead48740fef903594b