discovery: make checkbox_id invertable
We actually want to be able to get the services
back from the services checkbox_ids.
A hash function is designed to prevent just that.

Also note that the previous function was not injective,
as it would have conflated services of the form
("x_y", "z") and ("x", "y_z").

Change-Id: I837f9bd584ea2b75fd6d8a6493b0c3b0486bdcca