Remove BUILD_HOST.
Background: Calling execute_process() during CMake's configure stage
effectively forces a rebuild.  Doing this via a clever combination of
add_custom_command()/copy_if_changed/... might be a better solution, but
for now we just removed the parts which are not necessary or replace them
with predefined compiler macros. More to come...

Change-Id: I17a5a055971a5dbcf30e489a884c6a7ff61752cb