15903 FIX cisco_temperature, cisco_temperature_dom: Improved check state and summary if sensor status not ok
The cisco_temperature and cisco_temperature_dom checks currently only consider sensors with the state "ok". If a sensor becomes "unavailable" or "non-operational", the check becomes "UNKNOWN" with the summary "Item not found in monitoring data".
With this werk, the summary will now report the status of the sensor instead.
For the "unavailable" sensor status, the check state will still change to "UNKNOWN", for "non-operational" status it will be CRIT.

SUP-14650

Change-Id: I1f87fe96bf85ee48b4ddeca8a2e731dbeefa3c25