use query to bust caches
This might lower performance for people behind proxies, as the default
configuration for some proxies do not cache paths with query part. But
it simplifies the configuration of apache and other parts of the code.

In the long run we will add hashes to our filenames on disk and will
remove the _append_cache_busting_query for our internal files.

Change-Id: I4d41d528a1ab26af4cd8449c1fc0d00781fae735