Wrap access to graph_info (graph template registry)
Internally, we now use a frozen dataclass to represent graph templates.
All LazyStrings are localized when accessing the registry via the
wrapper.

Note that this is only an intermediate solution which has performance
drawbacks. However, this will be resolved when introducing a proper
registration mechanism for graph templates, which will absorb this
functionality.

CMK-13986

Change-Id: Ia15e4e82d59a286691a0e130e1d254400d3f7e24