Make HostAttributes a TypedDict
It's just a start. There are various possible improvements we can make
here, a few of them:

The host attributes may contain tags and custom attributes, which
can currently not be understood by the typing system.

There are also host- and folder exclusive attributes. To make the typing
understand the difference, we would have to have independent types for
host attributes and folder attributes.

We could also differentiate between effective attributes (where all
members are mandatory) and the explicit attributes (where all members
are optional).

Review all the dynamic attribute accesses and consider alternatives
which can be understood by the typing.

Change-Id: I8de3a27be60ab3ae6709b8015e79108c94cdf029