timeperiods: limiting recurring icalendar events to the current year
When importing an icalendar file, we now parse and extract recurring events.
However, the time horizon field allows the user to select up to 50 years
of recurring events.  At the moment we're limited to 75 recurring events
or the application will crash with an Internal Server Error.

There are changes on the horizon from the Engelbart team that may solve
this limitation but until then, we have to limit recurring events to
the current calendar year. This change implements that change. The
crash is still possible but the risk has been reduced greatly.

Change-Id: I7d12a0c7b977143341df30eebdf10b009d3b966d