Activate Changes redesign: Do not wait for whole step to finish
If one site takes longer to e.g. activate the remote changes, other sites should not be blocked in continuing their work.
We now maintain a list of active tasks that are periodically asked if they have finished in order to move on to the next step as soon as possible

CMK-13959

Change-Id: Ie43df6f09b5e0c6da2991df133eae63aaeca2288