14207 FIX BI state computation: Removed superfluous and expensive livestatus queries
It is possible to restrict the number of computed BI aggregation via filters.
If a filter was specified and no matching aggregation was found, a livestatus query over all aggregations was generated and sent to the core.
A huge response came back to the apache server, which dutifully parsed the response and threw it straight into the bin.

Change-Id: Id7345465af54110c3afcc7949ecf60b1fed202f3