CSE login redirects to cognito auth
When the login.py page is called the CSE edition will redirect to a page
handling the cognito oauth login. If we use the login.py to redirect we
can rely on the other logic in checkmk to forward a user to login.py
when a session is not active.

To achieve this we replace the login page in the saas edition. Once we
further refactor page registration we might be able to move this fully
into cse specific files and remove the edition check.

Change-Id: I2c6cd5dc1a391b3dcdd13853ad257d91216cfd07