Remove `as_autocheck_entry()` method
The method was added by feeea90ce1: "Handle API/check engine
impedance mismatch in API" because there was no better place
at that time. However, **this was a wrong move** as we do not
want to modify the check API!

Now that we have better layering, typing, and a few places
that alread handle most of the impedance mismatch, we move that
piece of logic over there.

CMK-12298

Change-Id: Id13dd4aa43407d2d6ad06002b47ff4cf23f22c8c