16029 Renamed MobileIron to Ivanti Neurons for MDM
MobileIron exists both as an on-premises and cloud version.
The included integration only applied for MobileIron Cloud,
which was not apparent by the name of the checks etc.

In the meantime, MobileIron was acquired and MobileIron Cloud was
renamed to Ivanti Neurons for MDM. This is now being properly
reflected in Checkmk.

Change-Id: Ie1ab465d0f7e05e558cabe17b364e142f7f82140