downtimes: client now has to specify the site_id when deleting downtimes
Downtime IDs are integers beginning from 1, both for the main site & for
remote sites.  This means that the same downtime ID can exist on more than
one site at the same time. This caused problems when calling the DELETE
endpoint as we weren't sure which ID the client wanted to delete.

To get around this problem, this werk introduces a mandatory site_id field
which the client has to specify when they want to delete a downtime.

SUP-14495

See werk #15966

Change-Id: Iadbc80f2d84b55551b650e3db7b11514fe258b79